Some unicode fixes. Some level 4 warning fixes. Ongoing. Added hotkey box to tool editor. Some hotkey code in tools.

+++ pnwtl/include/sslistctrl.h
@@ -100,7 +100,7 @@ class CSSListCtrl : public CListViewCtrl
 			{
 				memset(&pData->pItems[i], 0, sizeof(LVITEM));
 				pData->pItems[i].mask = LVIF_IMAGE | LVIF_PARAM | LVIF_STATE;
-				pData->pItems[i].stateMask = -1;
+				pData->pItems[i].stateMask = (UINT)-1;
 				pData->pItems[i].iItem = index;
 				pData->pItems[i].iSubItem = i;
